+++ src/mapnik_grid_view.cpp
@@ -263,29 +263,18 @@ NAN_METHOD(GridView::encodeSync)
     GridView* g = node::ObjectWrap::Unwrap<GridView>(args.Holder());
 
     // defaults
-    std::string format("utf");
     unsigned int resolution = 4;
     bool add_features = true;
 
-    // accept custom format
-    if (args.Length() >= 1){
-        if (!args[0]->IsString())
-        {
-            NanThrowTypeError("first arg, 'format' must be a string");
-            NanReturnUndefined();
-        }
-        format = TOSTR(args[0]);
-    }
-
     // options hash
-    if (args.Length() >= 2) {
-        if (!args[1]->IsObject())
+    if (args.Length() >= 1) {
+        if (!args[0]->IsObject())
         {
-            NanThrowTypeError("optional second arg must be an options object");
+            NanThrowTypeError("optional arg must be an options object");
             NanReturnUndefined();
         }
 
-        Local<Object> options = args[1].As<Object>();
+        Local<Object> options = args[0].As<Object>();
 
         if (options->Has(NanNew("resolution")))
         {
@@ -363,7 +352,6 @@ NAN_METHOD(GridView::encodeSync)
 typedef struct {
     uv_work_t request;
     GridView* g;
-    std::string format;
     bool error;
     std::string error_name;
     Persistent<Function> cb;
@@ -381,29 +369,18 @@ NAN_METHOD(GridView::encode)
     GridView* g = node::ObjectWrap::Unwrap<GridView>(args.Holder());
 
     // defaults
-    std::string format("utf");
     unsigned int resolution = 4;
     bool add_features = true;
 
-    // accept custom format
-    if (args.Length() >= 1){
-        if (!args[0]->IsString())
-        {
-            NanThrowTypeError("first arg, 'format' must be a string");
-            NanReturnUndefined();
-        }
-        format = TOSTR(args[0]);
-    }
-
     // options hash
-    if (args.Length() >= 2) {
-        if (!args[1]->IsObject())
+    if (args.Length() >= 1) {
+        if (!args[0]->IsObject())
         {
-            NanThrowTypeError("optional second arg must be an options object");
+            NanThrowTypeError("optional arg must be an options object");
             NanReturnUndefined();
         }
 
-        Local<Object> options = args[1].As<Object>();
+        Local<Object> options = args[0].As<Object>();
 
         if (options->Has(NanNew("resolution")))
         {
@@ -441,7 +418,6 @@ NAN_METHOD(GridView::encode)
     encode_grid_view_baton_t *closure = new encode_grid_view_baton_t();
     closure->request.data = closure;
     closure->g = g;
-    closure->format = format;
     closure->error = false;
     closure->resolution = resolution;
     closure->add_features = add_features;