+++ test/vector-tile.test.js
@@ -672,7 +672,7 @@ describe('mapnik.VectorTile ', function() {
 
         vtile.render(map, new mapnik.Grid(256, 256), {layer:0}, function(err, vtile_image) {
             if (err) throw err;
-            var utf = vtile_image.encodeSync('utf');
+            var utf = vtile_image.encodeSync();
             var expected_file = './test/data/vector_tile/tile0.expected.grid.json';
             var actual_file = './test/data/vector_tile/tile0.actual.grid.json';
             if (!existsSync(expected_file) || process.env.UPDATE) {